frameworks/tensorflow/tensorflow-neuron/tutorials/ssd300_demo/requirements.txt
neuron-cc
numpy>1.18.5
tensorflow-neuron<2.0
tensorflow-serving-api==1.15.0
torch>=1.0,<2.0
torchvision<1.0
matplotlib<4.0
Cython<0.29
pycocotools==2.0.1
tensorflow-serving-api==1.15.0
